server/requirements.txt
streamlit
streamlit_pdf_viewer
numpy
websockets
librosa
opencv-python
openai
# -- for pdf reading --
#pdfminer
#beautifulsoup4
python-doctr
python-doctr[torch]
pikepdf
pymysql
# -- api --
aiohttp
